# contract/openapi.yaml
openapi: 3.1.0

info:
  title: LopesTech Boilerplate API
  version: 1.0.0
  description: |
    Contrato canônico da família de boilerplates LopesTech.

    Este arquivo é versionado **idêntico** nos seis repositórios e é a fonte da verdade
    para dois consumidores:

    - a suíte de testes de contrato, que roda contra qualquer um dos seis backends;
    - os tipos TypeScript dos três frontends, gerados por `openapi-typescript`.

    Alterar este arquivo é uma mudança de contrato: exige atualizar os seis backends
    e rodar `make codegen` nos três frontends.
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: desenvolvimento local

tags:
  - name: health
    description: Sondas de liveness e readiness.
  - name: auth
    description: Registro, sessão e rotação de token.
  - name: users
    description: CRUD de usuários, protegido por Bearer.

security:
  - bearerAuth: []

paths:

  /api/v1/health:
    get:
      tags: [health]
      summary: Liveness
      description: Responde 200 enquanto o processo estiver de pé. Não toca no banco.
      operationId: healthLiveness
      security: []
      responses:
        "200":
          description: A aplicação está viva.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
              example: { status: ok, alive: true, ready: true }

  /api/v1/health/ready:
    get:
      tags: [health]
      summary: Readiness
      description: Verifica as dependências externas — hoje, a conexão com o banco.
      operationId: healthReadiness
      security: []
      responses:
        "200":
          description: A aplicação está pronta para receber tráfego.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
              example: { status: ok, alive: true, ready: true }
        "503":
          description: Alguma dependência está indisponível.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
              example: { status: unhealthy, alive: true, ready: false }

  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Registrar uma conta
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Conta criada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"

  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Abrir sessão
      description: |
        Devolve o access token no corpo — o cliente o mantém **em memória**.
        O refresh token vai em cookie `httpOnly; Secure; SameSite=Lax` e nunca
        aparece no corpo da resposta.
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Sessão aberta.
          headers:
            Set-Cookie:
              description: "refresh_token=<opaco>; HttpOnly; Secure; SameSite=Lax; Path=/api/v1/auth"
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationFailed"

  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotacionar o access token
      description: |
        Lê o refresh token do cookie, revoga-o e emite um par novo.
        Reuso de um refresh já consumido invalida toda a família de tokens.
      operationId: authRefresh
      security:
        - refreshCookie: []
      responses:
        "200":
          description: Token renovado.
          headers:
            Set-Cookie:
              description: Novo refresh token, rotacionado.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Encerrar sessão
      description: Revoga o refresh token e limpa o cookie.
      operationId: authLogout
      security:
        - refreshCookie: []
      responses:
        "204":
          description: Sessão encerrada.
          headers:
            Set-Cookie:
              description: Cookie expirado.
              schema: { type: string }

  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Perfil do portador do token
      operationId: authMe
      responses:
        "200":
          description: O usuário autenticado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/users:
    get:
      tags: [users]
      summary: Listar usuários
      operationId: usersList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          description: Busca parcial e insensível a maiúsculas por e-mail ou nome.
          required: false
          # Sem o padrão, o contrato promete aceitar um NUL — que não cabe num
          # `text` do Postgres e chega ao banco como 500.
          schema: { type: string, maxLength: 120, pattern: '^[^\x00-\x1f\x7f]{0,120}$' }
      responses:
        "200":
          description: Página de usuários.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    post:
      tags: [users]
      summary: Criar usuário
      operationId: usersCreate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: Usuário criado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"

  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      summary: Obter usuário
      operationId: usersGet
      responses:
        "200":
          description: O usuário.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    patch:
      tags: [users]
      summary: Atualizar usuário
      description: Atualização parcial. Campos ausentes permanecem inalterados.
      operationId: usersUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: Usuário atualizado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    delete:
      tags: [users]
      summary: Remover usuário
      operationId: usersDelete
      responses:
        "204":
          description: Usuário removido.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de vida curta, mantido em memória pelo cliente.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Token opaco, httpOnly, rotacionado a cada uso.

  parameters:
    Page:
      name: page
      in: query
      required: false
      # O teto existe porque sem ele o contrato promete aceitar um inteiro de
      # 25 dígitos, que estoura o int de todas as seis stacks. Uma página além
      # do fim já devolve lista vazia; pedir a milionésima não tem uso.
      schema: { type: integer, minimum: 1, maximum: 1000000, default: 1 }
    PerPage:
      name: per_page
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }

  responses:
    Unauthorized:
      description: Credenciais ausentes, inválidas ou expiradas.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://lopestech.dev/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: Credenciais inválidas.
    BadRequest:
      description: O corpo não pôde ser lido.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://lopestech.dev/errors/bad-request
            title: Bad Request
            status: 400
            detail: A requisição não pôde ser lida.
    NotFound:
      description: O recurso não existe.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://lopestech.dev/errors/not-found
            title: Not Found
            status: 404
            detail: Usuário não encontrado.
    Conflict:
      description: Violação de uma regra de unicidade.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://lopestech.dev/errors/conflict
            title: Conflict
            status: 409
            detail: Já existe uma conta com este e-mail.
    ValidationFailed:
      description: O corpo ou os parâmetros não passaram na validação.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblem" }
          example:
            type: https://lopestech.dev/errors/validation
            title: Validation failed
            status: 422
            detail: O corpo da requisição contém campos inválidos.
            errors:
              - field: email
                message: formato inválido

  schemas:

    Email:
      type: string
      format: email
      maxLength: 254
      # `format: email` sozinho admite endereços que os validadores das seis
      # stacks recusam — domínio sem ponto ("a@com"), rótulo iniciado por hífen,
      # TLD reservado pela RFC 2606. O padrão descreve a regra que de fato vale.
      #
      # É deliberadamente um subconjunto conservador: tudo que ele aceita é
      # aceito pela API. O contrário — contrato mais frouxo que a implementação —
      # é que faria o contrato mentir para quem o consome.
      pattern: '^[A-Za-z0-9_%+-]{1,20}(?:\.[A-Za-z0-9_%+-]{1,20}){0,3}@(?:[A-Za-z0-9](?:[A-Za-z0-9-]{0,20}[A-Za-z0-9])?\.){1,3}(?!(?:test|invalid|localhost)$)[A-Za-z]{2,10}$'
      examples: ["ana@exemplo.com"]

    Name:
      type: string
      minLength: 1
      maxLength: 120
      # Ao menos um caractere que não seja espaço, e nenhum caractere de
      # controle. Um nome só de espaços não é nome; e o byte NUL sequer cabe
      # num `text` do Postgres, então aceitá-lo viraria 500 na escrita.
      pattern: '^(?=\s*\S)[^\x00-\x1f\x7f]{1,120}$'
      examples: ["Ana Souza"]

    Secret:
      type: string
      format: password
      minLength: 12
      maxLength: 128
      # O comprimento é contado em pontos de código; caractere de controle é
      # recusado pelo mesmo motivo do nome.
      pattern: '^[^\x00-\x1f\x7f]{12,128}$'

    Problem:
      type: object
      description: Envelope de erro conforme a RFC 9457.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI que identifica o tipo do problema.
        title:
          type: string
          description: Resumo curto e estável do tipo do problema.
        status:
          type: integer
          description: O mesmo código HTTP da resposta.
        detail:
          type: string
          description: Explicação legível, específica desta ocorrência.
        instance:
          type: string
          format: uri
          description: URI da ocorrência específica.

    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              description: Um item por campo rejeitado.
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }

    HealthResponse:
      type: object
      required: [status, alive, ready]
      properties:
        status:
          type: string
          enum: [ok, unhealthy]
        alive: { type: boolean }
        ready: { type: boolean }

    User:
      type: object
      required: [id, email, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { $ref: "#/components/schemas/Email" }
        name: { $ref: "#/components/schemas/Name" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    RegisterRequest:
      type: object
      required: [email, name, password]
      # Campo desconhecido é erro, não algo a ignorar em silêncio:
      # é assim que um `passwrod` no cliente vira uma conta sem senha.
      additionalProperties: false
      properties:
        email: { $ref: "#/components/schemas/Email" }
        name: { $ref: "#/components/schemas/Name" }
        password:
          allOf: [{ $ref: "#/components/schemas/Secret" }]
          description: Mínimo de 12 caracteres. Hash com argon2id.

    LoginRequest:
      type: object
      required: [email, password]
      # Campo desconhecido é erro, não algo a ignorar em silêncio:
      # é assim que um `passwrod` no cliente vira uma conta sem senha.
      additionalProperties: false
      properties:
        email: { $ref: "#/components/schemas/Email" }
        password: { type: string, format: password }

    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Vida do access token, em segundos.

    UserCreate:
      type: object
      required: [email, name, password]
      # Campo desconhecido é erro, não algo a ignorar em silêncio:
      # é assim que um `passwrod` no cliente vira uma conta sem senha.
      additionalProperties: false
      properties:
        email: { $ref: "#/components/schemas/Email" }
        name: { $ref: "#/components/schemas/Name" }
        password: { $ref: "#/components/schemas/Secret" }

    UserUpdate:
      type: object
      minProperties: 1
      # Campo ausente permanece inalterado. Campo presente com `null` é erro:
      # não existe "apagar o e-mail". Distinguir as duas situações exige, em
      # várias das seis stacks, mais do que um tipo opcional.
      additionalProperties: false  # campo desconhecido também
      properties:
        email: { $ref: "#/components/schemas/Email" }
        name: { $ref: "#/components/schemas/Name" }

    PageMeta:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page: { type: integer, minimum: 1 }
        per_page: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        total_pages: { type: integer, minimum: 0 }

    UserPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        meta:
          $ref: "#/components/schemas/PageMeta"